refactor: Gathering 도메인 DB 관련 클래스 infrastructure 패키지에서 보관하도록 변경
