+++ momoim-core/src/main/java/com/triplem/momoim/core/domain/gathering/infrastructure/GatheringRepository.java
@@ -1,6 +1,10 @@
-package com.triplem.momoim.core.domain.gathering;
+package com.triplem.momoim.core.domain.gathering.infrastructure;
 
 import com.triplem.momoim.core.common.PaginationInformation;
+import com.triplem.momoim.core.domain.gathering.Gathering;
+import com.triplem.momoim.core.domain.gathering.GatheringDetail;
+import com.triplem.momoim.core.domain.gathering.GatheringPreview;
+import com.triplem.momoim.core.domain.gathering.GatheringSearchOption;
 import java.util.List;
 
 public interface GatheringRepository {
