+++ momoim-core/src/main/java/com/triplem/momoim/core/domain/gathering/infrastructure/GatheringTagConverter.java
@@ -1,4 +1,4 @@
-package com.triplem.momoim.core.domain.gathering;
+package com.triplem.momoim.core.domain.gathering.infrastructure;
 
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
