+++ momoim-core/src/main/java/com/triplem/momoim/core/domain/member/GatheringMemberRemover.java
@@ -1,7 +1,7 @@
 package com.triplem.momoim.core.domain.member;
 
 import com.triplem.momoim.core.domain.gathering.Gathering;
-import com.triplem.momoim.core.domain.gathering.GatheringRepository;
+import com.triplem.momoim.core.domain.gathering.infrastructure.GatheringRepository;
 import com.triplem.momoim.exception.BusinessException;
 import com.triplem.momoim.exception.ExceptionCode;
 import lombok.RequiredArgsConstructor;
