+++ momoim-core/src/test/java/com/triplem/momoim/core/domain/gathering/GatheringRepositoryTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import com.triplem.momoim.core.common.PaginationInformation;
 import com.triplem.momoim.core.common.SortOrder;
 import com.triplem.momoim.core.domain.gathering.enums.GatheringSortType;
+import com.triplem.momoim.core.domain.gathering.infrastructure.GatheringRepository;
 import com.triplem.momoim.core.domain.member.GatheringMember;
 import com.triplem.momoim.core.domain.member.GatheringMemberRepository;
 import java.util.List;
