+++ momoim-external-api/src/main/java/com/triplem/momoim/api/gathering/service/GatheringManagementService.java
@@ -1,8 +1,8 @@
 package com.triplem.momoim.api.gathering.service;
 
 import com.triplem.momoim.core.domain.gathering.Gathering;
-import com.triplem.momoim.core.domain.gathering.GatheringRepository;
 import com.triplem.momoim.core.domain.gathering.ModifyGathering;
+import com.triplem.momoim.core.domain.gathering.infrastructure.GatheringRepository;
 import com.triplem.momoim.core.domain.member.GatheringMemberAppender;
 import com.triplem.momoim.core.domain.member.GatheringMemberRemover;
 import com.triplem.momoim.exception.BusinessException;
