+++ momoim-external-api/src/test/java/com/triplem/momoim/api/gathering/GatheringJoinServiceTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.triplem.momoim.api.gathering.service.GatheringJoinService;
 import com.triplem.momoim.core.domain.gathering.Gathering;
-import com.triplem.momoim.core.domain.gathering.GatheringRepository;
+import com.triplem.momoim.core.domain.gathering.infrastructure.GatheringRepository;
 import com.triplem.momoim.core.domain.member.GatheringMember;
 import com.triplem.momoim.core.domain.member.GatheringMemberRepository;
 import org.junit.jupiter.api.DisplayName;
